requirements.txt
# Package dependencies
pandas
numpy
bs4
pyyaml
loguru
jupyter